code-play/backend/pyproject.toml
[project]
name = "code-play"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitignore-parser>=0.1.12",
    "httpx[socks]>=0.28.1",
    "jinja2>=3.1.6",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
